Extract `PollingSystem` abstraction on Native
