+++ core/native/src/main/scala/cats/effect/IOCompanionPlatform.scala
@@ -17,6 +17,9 @@
 package cats.effect
 
 import cats.effect.std.Console
+import cats.effect.unsafe.EventLoop
+
+import scala.reflect.ClassTag
 
 import java.time.Instant
 
@@ -62,4 +65,11 @@ private[effect] abstract class IOCompanionPlatform { this: IO.type =>
    */
   def readLine: IO[String] =
     Console[IO].readLine
+
+  def eventLoop[Poller](implicit ct: ClassTag[Poller]): IO[Option[EventLoop[Poller]]] =
+    IO.executionContext.map {
+      case loop: EventLoop[_] if ct.runtimeClass.isInstance(loop.poller()) =>
+        Some(loop.asInstanceOf[EventLoop[Poller]])
+      case _ => None
+    }
 }
