+++ core/native/src/main/scala/cats/effect/unsafe/EventLoop.scala
@@ -14,19 +14,13 @@
  * limitations under the License.
  */
 
-package cats.effect.unsafe
+package cats.effect
+package unsafe
 
-import scala.concurrent.duration._
+import scala.concurrent.ExecutionContext
 
-// JVM WSTP sets ExternalQueueTicks = 64 so we steal it here
-private[effect] object QueueExecutorScheduler extends PollingExecutorScheduler(64) {
-
-  def poll(timeout: Duration): Boolean = {
-    if (timeout != Duration.Zero && timeout.isFinite) {
-      val nanos = timeout.toNanos
-      Thread.sleep(nanos / 1000000, (nanos % 1000000).toInt)
-    }
-    false
-  }
+trait EventLoop[Poller] extends ExecutionContext {
+  
+  def poller(): Poller
 
 }
