+++ core/native/src/main/scala/cats/effect/unsafe/PollingExecutorScheduler.scala
@@ -17,65 +17,38 @@
 package cats.effect
 package unsafe
 
-import scala.concurrent.{ExecutionContext, ExecutionContextExecutor}
+import scala.concurrent.ExecutionContextExecutor
 import scala.concurrent.duration._
-import scala.scalanative.libc.errno
-import scala.scalanative.meta.LinktimeInfo
-import scala.scalanative.unsafe._
-import scala.util.control.NonFatal
 
-import java.util.{ArrayDeque, PriorityQueue}
-
+@deprecated("Use default runtime with a custom PollingSystem", "3.5.0")
 abstract class PollingExecutorScheduler(pollEvery: Int)
     extends ExecutionContextExecutor
-    with Scheduler {
+    with Scheduler { outer =>
 
-  private[this] var needsReschedule: Boolean = true
+  private[this] val loop = new EventLoopExecutorScheduler(
+    pollEvery,
+    new PollingSystem {
+      type Poller = outer.type
+      def makePoller(): Poller = outer
+      def close(poller: Poller): Unit = ()
+      def poll(poller: Poller, nanos: Long): Boolean =
+        if (nanos == -1) outer.poll(Duration.Inf) else outer.poll(nanos.nanos)
+    }
+  )
 
-  private[this] val executeQueue: ArrayDeque[Runnable] = new ArrayDeque
-  private[this] val sleepQueue: PriorityQueue[SleepTask] = new PriorityQueue
-
-  private[this] val noop: Runnable = () => ()
-
-  private[this] def scheduleIfNeeded(): Unit = if (needsReschedule) {
-    ExecutionContext.global.execute(() => loop())
-    needsReschedule = false
-  }
-
-  final def execute(runnable: Runnable): Unit = {
-    scheduleIfNeeded()
-    executeQueue.addLast(runnable)
-  }
+  final def execute(runnable: Runnable): Unit =
+    loop.execute(runnable)
 
   final def sleep(delay: FiniteDuration, task: Runnable): Runnable =
-    if (delay <= Duration.Zero) {
-      execute(task)
-      noop
-    } else {
-      scheduleIfNeeded()
-      val now = monotonicNanos()
-      val sleepTask = new SleepTask(now + delay.toNanos, task)
-      sleepQueue.offer(sleepTask)
-      sleepTask
-    }
+    loop.sleep(delay, task)
 
-  def reportFailure(t: Throwable): Unit = t.printStackTrace()
+  def reportFailure(t: Throwable): Unit = loop.reportFailure(t)
 
-  def nowMillis() = System.currentTimeMillis()
+  def nowMillis() = loop.nowMillis()
 
-  override def nowMicros(): Long =
-    if (LinktimeInfo.isFreeBSD || LinktimeInfo.isLinux || LinktimeInfo.isMac) {
-      import scala.scalanative.posix.time._
-      import scala.scalanative.posix.timeOps._
-      val ts = stackalloc[timespec]()
-      if (clock_gettime(CLOCK_REALTIME, ts) != 0)
-        throw new RuntimeException(s"clock_gettime: ${errno.errno}")
-      ts.tv_sec * 1000000 + ts.tv_nsec / 1000
-    } else {
-      super.nowMicros()
-    }
+  override def nowMicros(): Long = loop.nowMicros()
 
-  def monotonicNanos() = System.nanoTime()
+  def monotonicNanos() = loop.monotonicNanos()
 
   /**
    * @param timeout
@@ -89,66 +62,5 @@ abstract class PollingExecutorScheduler(pollEvery: Int)
    *   whether poll should be called again (i.e., there are more events to be polled)
    */
   protected def poll(timeout: Duration): Boolean
-
-  private[this] def loop(): Unit = {
-    needsReschedule = false
-
-    var continue = true
-
-    while (continue) {
-      // execute the timers
-      val now = monotonicNanos()
-      while (!sleepQueue.isEmpty() && sleepQueue.peek().at <= now) {
-        val task = sleepQueue.poll()
-        try task.runnable.run()
-        catch {
-          case t if NonFatal(t) => reportFailure(t)
-          case t: Throwable => IOFiber.onFatalFailure(t)
-        }
-      }
-
-      // do up to pollEvery tasks
-      var i = 0
-      while (i < pollEvery && !executeQueue.isEmpty()) {
-        val runnable = executeQueue.poll()
-        try runnable.run()
-        catch {
-          case t if NonFatal(t) => reportFailure(t)
-          case t: Throwable => IOFiber.onFatalFailure(t)
-        }
-        i += 1
-      }
-
-      // finally we poll
-      val timeout =
-        if (!executeQueue.isEmpty())
-          Duration.Zero
-        else if (!sleepQueue.isEmpty())
-          Math.max(sleepQueue.peek().at - monotonicNanos(), 0).nanos
-        else
-          Duration.Inf
-
-      val needsPoll = poll(timeout)
-
-      continue = needsPoll || !executeQueue.isEmpty() || !sleepQueue.isEmpty()
-    }
-
-    needsReschedule = true
-  }
-
-  private[this] final class SleepTask(
-      val at: Long,
-      val runnable: Runnable
-  ) extends Runnable
-      with Comparable[SleepTask] {
-
-    def run(): Unit = {
-      sleepQueue.remove(this)
-      ()
-    }
-
-    def compareTo(that: SleepTask): Int =
-      java.lang.Long.compare(this.at, that.at)
-  }
 
 }
